# openapi.yaml
openapi: 3.1.0
info:
  title: Kappa Simulation API
  description: API for running Kappa simulations using KaSim.
  version: 1.0.0
servers:
  - url: https://kappa-async.livecode.ch
    description: Kappa GPT Actions
paths:
  /run:
    post:
      operationId: run
      x-openai-isConsequential: false
      summary: Run Kappa simulation
      description: This endpoint accepts Kappa code, writes it to a file, executes the KaSim tool, and returns the simulation result.
      requestBody:
        description: JSON object containing the Kappa code and optional simulation parameters.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ka:
                  type: string
                  description: The Kappa code to simulate.
                  example: |
                    %agent: A(x)
                    A(x[.]) -> A(x[1]) @ 1.0
                l:
                  type: integer
                  description: Simulation limit
                  example: 100
                  default: 100
                p:
                  type: number
                  format: float
                  description: Plot period
                  example: 1.0
                  default: 1.0
              required:
                - ka
      responses:
        "200":
          description: Successful simulation execution
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
                    description: Output of the KaSim simulation
                  stdout:
                    type: string
                    description: The standard output of the KaSim simulation
                  stderr:
                    type: string
                    description: The standard error output of the KaSim simulation

  /run_async:
    post:
      operationId: runAsync
      x-openai-isConsequential: false
      summary: Run Kappa simulation asynchronously
      description: This endpoint accepts Kappa code, runs the KaSim tool in the background, and returns a unique key to retrieve the result later.
      requestBody:
        description: JSON object containing the Kappa code and optional simulation parameters.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ka:
                  type: string
                  description: The Kappa code to simulate.
                  example: |
                    %agent: A(x)
                    A(x[.]) -> A(x[1]) @ 1.0
                l:
                  type: integer
                  description: Simulation limit
                  example: 100
                  default: 100
                p:
                  type: number
                  format: float
                  description: Plot period
                  example: 1.0
                  default: 1.0
              required:
                - ka
      responses:
        "200":
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                    description: Unique key to retrieve the simulation result later
                  status:
                    type: string
                    example: "running"

  /run_async_result:
    post:
      operationId: getRunAsyncResult
      x-openai-isConsequential: false
      summary: Get Kappa simulation result for a specific task
      description: This endpoint retrieves the result of a previously initiated asynchronous Kappa simulation by providing a key in the request body.
      requestBody:
        description: JSON object containing the unique key to retrieve the result.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                key:
                  type: string
                  description: Unique key returned by the `/run_async` endpoint.
                  example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: Simulation result retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
                    description: Output of the KaSim simulation
                  stdout:
                    type: string
                    description: The standard output of the KaSim simulation
                  stderr:
                    type: string
                    description: The standard error output of the KaSim simulation
        "202":
          description: Simulation is still running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: Status of the simulation
                    example: "still running"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Task not found"